Average User Internet Session
The thought process behind solving this query was to separate the two entities that needed to be compared to find the differences between them. I used CTE's to gather the timestamp and dates of each user when they first loaded a screen on one cte and vise versa on the other. With this I was able to make a comparison between and find the average between both.